feat: PR #489 レビュー対応完了 - 品質向上とテストカバレッジ強化
- 300行制限違反修正: 3ファイルを機能別分割
  - performance.py: 495→52行（8モジュールに分割）
  - memory_analyzer.py: 479→252行（4モジュールに分割）
  - benchmark_runner.py: 422→120行（3モジュールに分割）

- テストカバレッジ強化: 4モジュールに包括的テスト追加
  - security_patterns.py: 17テストケース
  - error_display.py: 18テストケース
  - error_statistics.py: 15テストケース
  - recovery/manager.py: 15テストケース

- コード品質向上
  - 重複debug logger fallback定義削除（log_viewer.py）
  - 委譲パターンによる下位互換性維持
  - mypy strict mode完全準拠
